+++ History.md
@@ -3,3 +3,9 @@
 ==================
 
   * Initial release of `tcpproxy`.
+
+0.1.1 / 2012-03-18
+==================
+
+  * Smaller codebase.
+  * Added support for a listeningListener and a connectListener.